// src/app/api/fn/partners/partners-product-suppliers-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedProductSupplierList } from '../../models/paginated-product-supplier-list';

export interface PartnersProductSuppliersList$Params {

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Preferred links only.
 */
  preferred?: 'false' | 'true';

/**
 * Product id.
 */
  product?: number;

/**
 * Supplier id.
 */
  supplier?: number;
}

export function partnersProductSuppliersList(http: HttpClient, rootUrl: string, params?: PartnersProductSuppliersList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedProductSupplierList>> {
  const rb = new RequestBuilder(rootUrl, partnersProductSuppliersList.PATH, 'get');
  if (params) {
    rb.query('page', params.page, {});
    rb.query('preferred', params.preferred, {});
    rb.query('product', params.product, {});
    rb.query('supplier', params.supplier, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedProductSupplierList>;
    })
  );
}

partnersProductSuppliersList.PATH = '/api/partners/product-suppliers/';